Formatting the code to convention
