+++ src/SearchUsersPage/SearchUsersPage.css
@@ -1,0 +1,11 @@
+.note {
+	text-align: center;
+}
+
+.display-data {
+	margin: 0 1rem;
+}
+
+.user-data-container {
+	color: black;
+}
